Add files via upload
Explanation:

The code above defines a class TextEditor that represents a simple text editor. The constructor method sets up the GUI using the Tkinter library, creating a Text widget to display the text and a menu bar with options to create a new file, open an existing file, and save the current file.

The new_file(), open_file(), and save_file() methods implement the functionality of these options. new_file() simply clears the text widget and sets the filename to None. open_file() prompts the user to select a file using the filedialog.askopenfilename() function, reads the file contents into the text widget, and sets the filename. save_file() checks if a filename has been set, and if so, writes the contents of the text widget to the file. If no filename has been set, it prompts the user to select a filename using the filedialog.asksaveasfilename() function